requirements.txt
Appium-Python-Client
opencv-python
pillow
pytest
scikit-image
